Outline: Move OutlineExplorerWidget to its own module
