+++ spyder/plugins/outlineexplorer/plugin.py
@@ -13,7 +13,7 @@ from qtpy.QtCore import Slot
 from spyder.api.plugin_registration.decorators import on_plugin_available
 from spyder.api.translations import get_translation
 from spyder.api.plugins import SpyderDockablePlugin, Plugins
-from spyder.plugins.outlineexplorer.widgets import OutlineExplorerWidget
+from spyder.plugins.outlineexplorer.main_widget import OutlineExplorerWidget
 
 # Localization
 _ = get_translation('spyder')
